plane db changed to zip

+++ JAERO/mskdemodulator.cpp
@@ -310,7 +310,7 @@ qint64 MskDemodulator::writeData(const char *data, qint64 len)
     bool sendscatterpoints=false;
 
     const short *ptr = reinterpret_cast<const short *>(data);
-    for(int i=0;i<len/sizeof(short);i++)
+    for(int i=0;i<(int)(len/sizeof(short));i++)
     {
 
         double dval=((double)(*ptr))/32768.0;
